Fix enum property drawer

+++ GloryEngine/Editor/GloryEditor/EditorUI.h
@@ -37,12 +37,12 @@ namespace Glory::Editor
 
 		static GLORY_EDITOR_API bool InputDropdown(std::string_view label, const std::vector<std::string_view>& options, size_t* index, std::string_view value);
 		static GLORY_EDITOR_API bool InputDropdown(std::string_view label, const std::vector<std::string_view>& options, std::vector<bool>& selected, std::string_view value);
-		static GLORY_EDITOR_API bool InputEnum(std::string_view label, uint32_t typeHash, size_t* value, std::vector<size_t> excludeValues = std::vector<size_t>());
+		static GLORY_EDITOR_API bool InputEnum(std::string_view label, uint32_t typeHash, uint32_t* value, std::vector<uint32_t> excludeValues = std::vector<uint32_t>());
 
 		template<typename T>
-		static bool InputEnum(std::string_view label, T* value, std::vector<size_t> excludeValues = std::vector<size_t>())
+		static bool InputEnum(std::string_view label, T* value, std::vector<uint32_t> excludeValues = std::vector<uint32_t>())
 		{
-			return InputEnum(label, ResourceType::GetHash<T>(), (size_t*)value, excludeValues);
+			return InputEnum(label, ResourceType::GetHash<T>(), (uint32_t*)value, excludeValues);
 		}
 
 		static GLORY_EDITOR_API bool InputLayerMask(std::string_view label, LayerMask* data);
